Rename module folder to src
